backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
resend
email-validator
